Refactor into monitoring module
